Refactor the codes and add distance calculation static class.
